Integrate user input into main loop.
This is a change to the main loop. The loop is now rate limited by the
generation of events from the window. Previously camera did this, but camera
has been removed. If the window is connected to the same OpenGL as the module
we write to render things to the screen, the window and that module don't need
to know about each other.

All life implementers will get access to the current loop event.